tests/wpt/web-platform-tests/html/semantics/forms/the-input-element/pattern_attribute.html

<!DOCTYPE html>
<title>pattern attribute</title>
<meta name=viewport content="width=device-width">
<link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
<link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#attr-input-pattern">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h1><code>pattern</code> attribute</h1>
<div style="display: none">
  <input pattern="[a-z]{3}" value="abcd" title="three letters max">
</div>
<div id="log"></div>
<script>
  test(function() {
    const input = document.querySelector("input");

    assert_idl_attribute(input, "pattern");
    assert_equals(input.pattern, "[a-z]{3}");

    assert_inherits(input, "validity");
    assert_false(input.validity.valid);

    assert_true(input.matches(":invalid"));
  }, "pattern attribute support on input element");
</script>
